Plotting update (#105)
Update things to work with the latest versions of julia and makie. In addition, the plotting infrastructure has been revamped to make it trivial to add custom plots both through the package and independently by users. Plotting now supports saving of images.

One thing I experimented with, but did not implement in the end, was plotting through individual interations. Hacked a setup to do that and it can be really powerful, but one would need to properly think out how to do it cleanly.

* Prototype of adjusted live plotting

* Swith GLMakie dependency for just Makie

* Update julia version in CI

* implemented KippenLine

* Improve HRPlot

* Updated Project.toml of docs

* Added abundance plot

Additionally, small fixes and custom theme

* add missing Abundances.jl file

* Fix to doc plots and small IO fix

* Adjustments to theming and abundance labels

* Function parity for TRho plot

* Clean up NuclearBurning.jl

* Simply load the empty History and Profile files

These are empty, need to redo them

* Allow KippLine to use mass/radius model_number/time

* Fixes to KippLine

* Fix in StellarEvolution

* Add options to save plot

* Finish making pretty labels

* Option renaming

* Implemented History and Profile plots

Added all implemented plots to NuclearBurning.jl to highlight them

* Small updates to NuclearBurning

* Fix NullPlotter, make output options part of model

* Fix OneZone

* Small fix to labels for abundances

* update prject tomls

* typos

* make dummy lines! call plot length 1 arrays

---------

Co-authored-by: matthiasfabry <matthias.fabry@villanova.edu>